refactor: rename summarizer to reporter
This changes the nomenclature of the interface for reporting policy
statuses.

BREAKING CHANGE: This renames the `--summary` flag to `--reporter`.
